// src/app/home.interface.ts
export interface AllproductsResponse {
  results: number
  metadata: Metadata
  data: AllproductsData[]
}

export interface Metadata {
  currentPage: number
  numberOfPages: number
  limit: number
  nextPage: number
}

export interface AllproductsData {
  sold?: number
  images: string[]
  subcategory: Subcategory[]
  ratingsQuantity: number
  _id: string
  title: string
  slug: string
  description: string
  quantity: number
  price: number
  imageCover: string
  category: Category
  brand: Brand
  ratingsAverage: number
  createdAt: string
  updatedAt: string
  id: string
  priceAfterDiscount?: number
  availableColors?: any[]
}

export interface Subcategory {
  _id: string
  name: string
  slug: string
  category: string
}

export interface Category {
  _id: string
  name: string
  slug: string
  image: string
}

export interface Brand {
  _id: string
  name: string
  slug: string
  image: string
}

export interface AllCategoriesResponse {
  results: number
  metadata: Metadata
  data: CategoriesData[]
}

export interface CategoriesData {
  _id: string
  name: string
  slug: string
  image: string
  createdAt: string
  updatedAt: string
}